pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-demo"
version = "0.1.0"
description = "A RAG demo using LlamaIndex"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "RAG Demo" },
]
dependencies = [
    # LlamaIndex Core
    "llama-index-core>=0.12.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-llms-ollama>=0.3.0",
    "llama-index-embeddings-huggingface>=0.3.0",
    "llama-index-llms-deepseek>=0.1.2",
    "llama-index-llms-openai-like>=0.4.0",
    
    # ML/AI Libraries
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    
    # Web API Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    
    # Document Processing
    "pypdf>=3.0.0",
    "pdfplumber>=0.9.0",
    "python-magic>=0.4.0",
    
    # Utilities
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
    
    # Vector storage
    "elasticsearch>=8.12.0",
    "llama-index-vector-stores-elasticsearch>=0.2.0",
    
    # CLI Interface (kept for backward compatibility)
    "rich>=13.0.0",
]

[project.scripts]
rag-demo = "kb.main:main"
rag-web = "kb.api.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "httpx>=0.25.0",  # For API testing
]

[project.optional-dependencies]
web = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "psutil>=5.9.0",
]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100 

[tool.hatch.build.targets.wheel]
packages = ["src/kb"]

[tool.hatch.build.targets.sdist]
include = [
    "src/kb",
    "README.md",
    "LICENSE*",
    "pyproject.toml",
]
